Expose branch env variable to the client-side
In order to properly detect a produuction deployment and set links
correctly when in production, the appropriate env variable needs
to be exposed in the next.config.js file.

The commit also renames the variable to use the `VERCEL_`, since
documentation on the legacy `NOW_` prefix is now pretty much
discontinued.