Use feature branch prefixes temporarily
This reverts temporatily the long term solution in the
previous commit to accomodate for a limitation: the Services API
doesn't fallback to "master-services.jenkins..." if a given
feature branch doesn't exist.

Forcing client developers to deploy API branches just for testing
preview environments is not practical. The long term solution
requires discussion, but this temporary solution allows for
API developers to validate Steer against a specific API
feature branch given it starts with "sqa-".